+++ CHANGELOG.md
@@ -9,12 +9,12 @@
 ### Bug-fixes
 
 - [x] [#488 - Failed to copy icon folder](https://github.com/scriptPilot/app-framework/issues/488)
+- [x] [#541 - firebase instead of devFirebase config used for npm run ios/android](https://github.com/scriptPilot/app-framework/issues/541)
 
 ### Improvements
 
 - [x] [#551 - Allow "npm run android" without configured playStoreId](https://github.com/scriptPilot/app-framework/issues/551)
 - [x] [#552 - Allow "npm run ios" without configured appStoreId](https://github.com/scriptPilot/app-framework/issues/552)
-
 
 ## Version 1.6.52